projects/Telegram_bot_for_tracking_cryp/source_code/requirements.txt
```
APScheduler==3.10.1
psycopg2-binary==2.9.8
python-telegram-bot==20.0
requests==2.31.0
SQLAlchemy==2.0.21
```